refactor: 커피챗 단계 설명 툴팁 컴포넌트 이동
