--- app/[locale]/(main)/coffeechat/components/CoffeeChatCount/index.ts
@@ -1,1 +1,0 @@
-export * from "./CoffeeChatCount";
