+++ app/[locale]/(main)/coffeechat/components/CoffeeChatSteps/index.ts
@@ -1,0 +1,1 @@
+export * from "./CoffeeChatSteps";
